requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-huggingface

google-generativeai
python-dotenv

sentence-transformers
faiss-cpu

pydantic

streamlit
numpy
scikit-learn
